+++ opensearchapi/api.indices.delete_datastream.go
@@ -7,23 +7,6 @@
 // Modifications Copyright OpenSearch Contributors. See
 // GitHub history for details.
 
-// Licensed to Elasticsearch B.V. under one or more contributor
-// license agreements. See the NOTICE file distributed with
-// this work for additional information regarding copyright
-// ownership. Elasticsearch B.V. licenses this file to you under
-// the Apache License, Version 2.0 (the "License"); you may
-// not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//    http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
-
 package opensearchapi
 
 import (
@@ -46,12 +29,9 @@ func newIndicesDeleteDataStreamFunc(t Transport) IndicesDeleteDataStream {
 // ----- API Definition -------------------------------------------------------
 
 // IndicesDeleteDataStream deletes the backing indices of a data stream and then deletes the data stream itself.
-//
-//
 type IndicesDeleteDataStream func(name string, o ...func(*IndicesDeleteDataStreamRequest)) (*Response, error)
 
 // IndicesDeleteDataStreamRequest configures the Data Stream Delete Template API request.
-//
 type IndicesDeleteDataStreamRequest struct {
 	Name string
 
@@ -69,7 +49,6 @@ type IndicesDeleteDataStreamRequest struct {
 }
 
 // Do execute the request and returns response or error.
-//
 func (r IndicesDeleteDataStreamRequest) Do(ctx context.Context, transport Transport) (*Response, error) {
 	var (
 		method string
@@ -80,9 +59,7 @@ func (r IndicesDeleteDataStreamRequest) Do(ctx context.Context, transport Transp
 	method = "DELETE"
 
 	path.Grow(1 + len("_data_stream") + 1 + len(r.Name))
-	path.WriteString("/")
-	path.WriteString("_data_stream")
-	path.WriteString("/")
+	path.WriteString("/_data_stream/")
 	path.WriteString(r.Name)
 
 	params = make(map[string]string)
@@ -155,7 +132,6 @@ func (r IndicesDeleteDataStreamRequest) Do(ctx context.Context, transport Transp
 }
 
 // WithContext sets the request context.
-//
 func (f IndicesDeleteDataStream) WithContext(v context.Context) func(*IndicesDeleteDataStreamRequest) {
 	return func(r *IndicesDeleteDataStreamRequest) {
 		r.ctx = v
@@ -163,7 +139,6 @@ func (f IndicesDeleteDataStream) WithContext(v context.Context) func(*IndicesDel
 }
 
 // WithClusterManagerTimeout - explicit operation timeout for connection to cluster-manager node.
-//
 func (f IndicesDeleteDataStream) WithClusterManagerTimeout(v time.Duration) func(*IndicesDeleteDataStreamRequest) {
 	return func(r *IndicesDeleteDataStreamRequest) {
 		r.ClusterManagerTimeout = v
@@ -171,7 +146,6 @@ func (f IndicesDeleteDataStream) WithClusterManagerTimeout(v time.Duration) func
 }
 
 // WithTimeout - explicit operation timeout.
-//
 func (f IndicesDeleteDataStream) WithTimeout(v time.Duration) func(*IndicesDeleteDataStreamRequest) {
 	return func(r *IndicesDeleteDataStreamRequest) {
 		r.Timeout = v
@@ -179,7 +153,6 @@ func (f IndicesDeleteDataStream) WithTimeout(v time.Duration) func(*IndicesDelet
 }
 
 // WithPretty makes the response body pretty-printed.
-//
 func (f IndicesDeleteDataStream) WithPretty() func(*IndicesDeleteDataStreamRequest) {
 	return func(r *IndicesDeleteDataStreamRequest) {
 		r.Pretty = true
@@ -187,7 +160,6 @@ func (f IndicesDeleteDataStream) WithPretty() func(*IndicesDeleteDataStreamReque
 }
 
 // WithHuman makes statistical values human-readable.
-//
 func (f IndicesDeleteDataStream) WithHuman() func(*IndicesDeleteDataStreamRequest) {
 	return func(r *IndicesDeleteDataStreamRequest) {
 		r.Human = true
@@ -195,7 +167,6 @@ func (f IndicesDeleteDataStream) WithHuman() func(*IndicesDeleteDataStreamReques
 }
 
 // WithErrorTrace includes the stack trace for errors in the response body.
-//
 func (f IndicesDeleteDataStream) WithErrorTrace() func(*IndicesDeleteDataStreamRequest) {
 	return func(r *IndicesDeleteDataStreamRequest) {
 		r.ErrorTrace = true
@@ -203,7 +174,6 @@ func (f IndicesDeleteDataStream) WithErrorTrace() func(*IndicesDeleteDataStreamR
 }
 
 // WithFilterPath filters the properties of the response body.
-//
 func (f IndicesDeleteDataStream) WithFilterPath(v ...string) func(*IndicesDeleteDataStreamRequest) {
 	return func(r *IndicesDeleteDataStreamRequest) {
 		r.FilterPath = v
@@ -211,7 +181,6 @@ func (f IndicesDeleteDataStream) WithFilterPath(v ...string) func(*IndicesDelete
 }
 
 // WithHeader adds the headers to the HTTP request.
-//
 func (f IndicesDeleteDataStream) WithHeader(h map[string]string) func(*IndicesDeleteDataStreamRequest) {
 	return func(r *IndicesDeleteDataStreamRequest) {
 		if r.Header == nil {
@@ -224,7 +193,6 @@ func (f IndicesDeleteDataStream) WithHeader(h map[string]string) func(*IndicesDe
 }
 
 // WithOpaqueID adds the X-Opaque-Id header to the HTTP request.
-//
 func (f IndicesDeleteDataStream) WithOpaqueID(s string) func(*IndicesDeleteDataStreamRequest) {
 	return func(r *IndicesDeleteDataStreamRequest) {
 		if r.Header == nil {